+++ src/config/config.h
@@ -92,6 +92,7 @@ class Config : public QObject {
     Q_PROPERTY(bool showBatteryEveryWhere READ getShowBatteryEveryWhere WRITE setShowBatteryEveryWhere NOTIFY showBatteryEveryWhereChanged)
 
     Q_PROPERTY(bool enableActivityBar READ getEnableActivityBar WRITE setEnableActivityBar NOTIFY enableActivityBarChanged)
+    Q_PROPERTY(bool openActivityOnApiStart READ getOpenActivityOnApiStart WRITE setOpenActivityOnApiStart NOTIFY openActivityOnApiStartChanged)
     Q_PROPERTY(bool fillMediaArtwork READ getFillMediaArtwork WRITE setFillMediaArtwork NOTIFY fillMediaArtworkChanged)
     Q_PROPERTY(bool mediaCoverflowDefault READ getMediaCoverflowDefault WRITE setMediaCoverflowDefault NOTIFY mediaCoverflowDefaultChanged)
 
@@ -174,6 +175,9 @@ class Config : public QObject {
 
     bool getEnableActivityBar();
     void setEnableActivityBar(bool value);
+
+    bool getOpenActivityOnApiStart();
+    void setOpenActivityOnApiStart(bool value);
 
     bool getFillMediaArtwork();
     void setFillMediaArtwork(bool value);
@@ -318,6 +322,7 @@ class Config : public QObject {
     void showBatteryPercentageChanged();
     void showBatteryEveryWhereChanged();
     void enableActivityBarChanged();
+    void openActivityOnApiStartChanged();
     void fillMediaArtworkChanged();
     void mediaCoverflowDefaultChanged();
     void resumeTimeoutWindowSecChanged(int value);